refactor(claim-gates): split claim-drift-engine into modules
Break the 3487-line claim-drift-engine into claim-drift/{state,metrics,scanners,run}
with a thin public barrel. Preserves historical import path and exports.
Fleet-redundancy C12 length residual.
